refactor(cli): unify console output into display module
Replace console.rs with a display/ module — one Describe impl per entity
rendering an aligned block, a centralized theme, and per-entity files.
Merge the train callbacks (ConsoleMonitor, ModelSaver) into callbacks.rs.
